Coppel-page/coppel-backend/coppel-back/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.3.0
setuptools==69.5.1
sqlparse==0.4.4
tzdata==2024.1
wheel==0.43.0
